Refactor: Theme sources moved to ui/res/.
