+++ lib/ui/screen/sight_details_screen.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:places/domain/sight.dart';
 import 'package:places/ui/image_loading_progress.dart';
-import 'package:places/ui/screen/theme/text_kit.dart';
-import 'package:places/ui/screen/theme/themes.dart';
+import 'package:places/ui/res/text_kit.dart';
+import 'package:places/ui/res/themes.dart';
 import 'package:places/ui/screen/widget/my_app_bar.dart';
 import 'package:places/ui/svg_icon.dart';
