+++ lib/ui/screen/widget/sight_card.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:places/domain/sight.dart';
 import 'package:places/ui/image_loader.dart';
 import 'package:places/ui/screen/sight_details_screen.dart';
-import 'package:places/ui/screen/theme/themes.dart';
+import 'package:places/ui/res/themes.dart';
 
 class SightCard extends StatefulWidget {
   final Sight sight;
